Added support for mTLS
This commit adds support for mutual authentication via TLS (i.e. mTLS).

The server now has an option (`--tls_client_ca`) to allow the user to
specify the certificate with which connecting client certificates need
to be signed.

The client now has new options (`--tls_certificate` and
`tls_private_key`) to specify which key and certificate can be used when
connecting to the server.